simply add page for adding lesson
