+++ AndroidManifest.xml
@@ -23,6 +23,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".activity.AddLessonActivity"
+            android:label="@string/title_activity_add_lesson" >
+        </activity>
     </application>
 
 </manifest>
