+++ res/layout/activity_add_lesson.xml
@@ -1,0 +1,40 @@
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:actbar="http://schemas.android.com/apk/res/xziar.mylesson"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    tools:context="${relativePackage}.${activityClass}" >
+    
+    <xziar.mylesson.view.actionbar.SimpleActionBar
+        android:id="@+id/actbar"
+        android:layout_width="match_parent"
+        android:layout_height="32dp"
+        android:layout_margin="5dp">
+
+        <xziar.mylesson.view.actionbar.ActionBarButton 
+            android:layout_width="wrap_content"
+        	android:layout_height="match_parent"
+        	actbar:Align="left"
+        	actbar:img="@drawable/icon_close"
+        	android:onClick="onBtnClose"
+            />
+
+        <xziar.mylesson.view.actionbar.ActionBarButton 
+            android:layout_width="wrap_content"
+        	android:layout_height="match_parent"
+        	actbar:Align="right"
+        	actbar:img="@drawable/icon_yes"
+        	android:onClick="onBtnYes"
+            />
+    	
+    </xziar.mylesson.view.actionbar.SimpleActionBar>
+    
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical" >
+        
+    </LinearLayout>
+    
+</LinearLayout>
