+++ res/values/strings.xml
@@ -3,5 +3,6 @@
 
     <string name="app_name">MyLesson</string>
     <string name="hello_world">Hello world!</string>
+    <string name="title_activity_add_lesson">AddLessonActivity</string>
 
 </resources>
